Initial upload
A simple Python (3.4) program that emulates physical flash cards.  The
"decks" directory contains .csv files with information on the flash
cards.  The first column is the "front" of the card, the second column
is the "back" of the card, and the third column is "more" information.
Variable "first_side" determines if the each card should be shown front
side first, back side first, or randomly chosen between the front and
back sides.  More information is always shown after the front and back
sides have both been revealed.